live/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-osc>=1.8.3
pyyaml>=6.0
httpx>=0.25.0
websockets>=12.0
